requirements.txt
Flask
torch
timm
fastcore
torchvision
Pillow
flask_cors